Implemented new design
-altered the repo and class system to represent new design
- no functionality changes
