+++ classes/colours.py
@@ -1,0 +1,15 @@
+class Colours:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    BROWN = "\033[0;33m"
+    YELLOW = "\033[1;33m"
+    PURPLE = "\033[0;35m"
+    LIGHT_CYAN = "\033[1;36m"
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+
+
