+++ classes/inventory.py
@@ -1,7 +1,7 @@
 class Item:
-    def __init__(self, name, type, description, prop):
+    def __init__(self, name, form, description, prop):
         self.name = name
-        self.type = type
+        self.form = form
         self.description = description
         self.prop = prop
